TLS verification struct now reside top-level in the `commons` module
